Custom Frontend/requirements.txt
Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.0
